:recycle: refactored LSL functions into lsl_stream.py
